#70, #73: Various updates and re-factoring
image_handler.php updated for #70; the uninstall option is no longer
displayed if the admin doesn't have access.

Other changes: Update the image-upload handling to supply a list of
"acceptable" file extensions, check that the to-be-updated image name
actually "fits" into the database field, using "pathinfo" to gather the
directory/name/extension information for images.

When searching for additional images, use the storefront algorithm (the
match-parameters are different if the file is in the root /images
extension instead of a subdirectory).